Fix manually setting the flagship for OAuth clients
