+++ web/auth/flagship.go
@@ -1,10 +1,13 @@
 package auth
 
 import (
+	"encoding/json"
 	"net/http"
 	"strings"
 	"time"
 
+	"github.com/cozy/cozy-stack/model/oauth"
+	"github.com/cozy/cozy-stack/pkg/limits"
 	"github.com/cozy/cozy-stack/web/middlewares"
 	"github.com/labstack/echo/v4"
 )
@@ -41,4 +44,74 @@ func CreateSessionCode(c echo.Context) error {
 	return c.JSON(http.StatusCreated, echo.Map{
 		"session_code": code,
 	})
+}
+
+func postChallenge(c echo.Context) error {
+	inst := middlewares.GetInstance(c)
+	err := limits.CheckRateLimit(inst, limits.OAuthClientType)
+	if limits.IsLimitReachedOrExceeded(err) {
+		return echo.NewHTTPError(http.StatusNotFound, "Not found")
+	}
+	client := c.Get("client").(*oauth.Client)
+	nonce, err := client.CreateChallenge(inst)
+	if err != nil {
+		return c.JSON(http.StatusInternalServerError, echo.Map{"error": err.Error()})
+	}
+	return c.JSON(http.StatusCreated, echo.Map{"nonce": nonce})
+}
+
+func postAttestation(c echo.Context) error {
+	inst := middlewares.GetInstance(c)
+	client, err := oauth.FindClient(inst, c.Param("client-id"))
+	if err != nil {
+		return c.JSON(http.StatusNotFound, echo.Map{
+			"error": "Client not found",
+		})
+	}
+	var data oauth.AttestationRequest
+	if err := json.NewDecoder(c.Request().Body).Decode(&data); err != nil {
+		return c.JSON(http.StatusBadRequest, echo.Map{
+			"error": err.Error(),
+		})
+	}
+	if err := client.Attest(inst, data); err != nil {
+		inst.Logger().Infof("Cannot attest %s client: %s", client.ID(), err.Error())
+		return c.JSON(http.StatusBadRequest, echo.Map{
+			"error": err.Error(),
+		})
+	}
+	return c.NoContent(http.StatusNoContent)
+}
+
+func confirmFlagship(c echo.Context) error {
+	inst := middlewares.GetInstance(c)
+	client, err := oauth.FindClient(inst, c.Param("client-id"))
+	if err != nil {
+		return c.JSON(http.StatusNotFound, echo.Map{
+			"error": "Client not found",
+		})
+	}
+
+	err = limits.CheckRateLimit(inst, limits.ConfirmFlagshipType)
+	if limits.IsLimitReachedOrExceeded(err) {
+		return c.JSON(http.StatusUnauthorized, echo.Map{
+			"error": inst.Translate("Confirm Flagship Invalid code"),
+		})
+	}
+
+	clientID := c.Param("client-id")
+	code := c.FormValue("code")
+	token := []byte(c.FormValue("token"))
+	if ok := oauth.CheckFlagshipCode(inst, clientID, token, code); !ok {
+		return c.JSON(http.StatusUnauthorized, echo.Map{
+			"error": inst.Translate("Confirm Flagship Invalid code"),
+		})
+	}
+
+	if err := client.SetFlagship(inst); err != nil {
+		return c.JSON(http.StatusInternalServerError, echo.Map{
+			"error": err.Error,
+		})
+	}
+	return c.NoContent(http.StatusNoContent)
 }
